Paystack changes

+++ src/main/java/com/africa/semiclon/capStoneProject/dtos/request/PaymentRequest.java
@@ -1,5 +1,9 @@
 package com.africa.semiclon.capStoneProject.dtos.request;
 
+import com.africa.semiclon.capStoneProject.constants.CustomBigDecimalDeserializer;
+import com.africa.semiclon.capStoneProject.constants.CustomBigDecimalSerializer;
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -9,5 +13,5 @@ import java.math.BigDecimal;
 @Setter
 public class PaymentRequest {
     private Long userId;
-    private BigDecimal amount;
+    private int amount;
 }
